Basic solar support.

+++ Adventure/Geom.hpp
@@ -59,7 +59,7 @@ namespace Ad
     Geom (const Attrib* attribs, uint count, uint index_buffer, uint indx_count, uint index_type);
 
     Geom (std::initializer_list <Attrib> attribs, uint index_buffer, uint index_count, uint index_type) :
-      Geom (attribs.begin (), attribs.size (), index_buffer, index_count, index_type)
+      Geom (attribs.begin (), (uint) attribs.size (), index_buffer, index_count, index_type)
     { }
 
     ~Geom ()
